referral_bot/requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
flask==3.0.0
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0
apscheduler==3.10.4
flask-limiter==3.5.0
cryptography==41.0.7
httpx==0.25.2
